requirements.txt
langgraph
python-dotenv
pandas
requests
tavily-python
gradio

fastapi
uvicorn[standard]
jinja2
